added changes for deploy

+++ .next/dev/server/edge/chunks/_e36d360d._.js
@@ -2646,7 +2646,7 @@ var MiddlewareSpan = /*#__PURE__*/ function(MiddlewareSpan) {
     MiddlewareSpan["execute"] = "Middleware.execute";
     return MiddlewareSpan;
 }(MiddlewareSpan || {});
-const NextVanillaSpanAllowlist = [
+const NextVanillaSpanAllowlist = new Set([
     "Middleware.execute",
     "BaseServer.handleRequest",
     "Render.getServerSideProps",
@@ -2663,12 +2663,12 @@ const NextVanillaSpanAllowlist = [
     "NextNodeServer.getLayoutOrPageModule",
     "NextNodeServer.startResponse",
     "NextNodeServer.clientComponentLoading"
-];
-const LogSpanAllowList = [
+]);
+const LogSpanAllowList = new Set([
     "NextNodeServer.findPageComponents",
     "NextNodeServer.createComponentTree",
     "NextNodeServer.clientComponentLoading"
-];
+]);
 ;
  //# sourceMappingURL=constants.js.map
 }),
@@ -4192,6 +4192,7 @@ var __TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist
 var __TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$shared$2f$lib$2f$is$2d$thenable$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__ = __turbopack_context__.i("[project]/node_modules/next/dist/esm/shared/lib/is-thenable.js [middleware-edge] (ecmascript)");
 ;
 ;
+const NEXT_OTEL_PERFORMANCE_PREFIX = process.env.NEXT_OTEL_PERFORMANCE_PREFIX;
 let api;
 // we want to allow users to use their own version of @opentelemetry/api if they
 // want to, so we try to require it first, and if it fails we fall back to the
@@ -4271,7 +4272,6 @@ class NextTracerImpl {
         return context.with(remoteContext, fn);
     }
     trace(...args) {
-        var _trace_getSpanContext;
         const [type, fnOrOptions, fnOrEmpty] = args;
         // coerce options form overload
         const { fn, options } = typeof fnOrOptions === 'function' ? {
@@ -4284,18 +4284,20 @@ class NextTracerImpl {
             }
         };
         const spanName = options.spanName ?? type;
-        if (!__TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$server$2f$lib$2f$trace$2f$constants$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["NextVanillaSpanAllowlist"].includes(type) && process.env.NEXT_OTEL_VERBOSE !== '1' || options.hideSpan) {
+        if (!__TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$server$2f$lib$2f$trace$2f$constants$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["NextVanillaSpanAllowlist"].has(type) && process.env.NEXT_OTEL_VERBOSE !== '1' || options.hideSpan) {
             return fn();
         }
         // Trying to get active scoped span to assign parent. If option specifies parent span manually, will try to use it.
         let spanContext = this.getSpanContext((options == null ? void 0 : options.parentSpan) ?? this.getActiveScopeSpan());
-        let isRootSpan = false;
         if (!spanContext) {
             spanContext = (context == null ? void 0 : context.active()) ?? ROOT_CONTEXT;
-            isRootSpan = true;
-        } else if ((_trace_getSpanContext = trace.getSpanContext(spanContext)) == null ? void 0 : _trace_getSpanContext.isRemote) {
-            isRootSpan = true;
         }
+        // Check if there's already a root span in the store for this trace
+        // We are intentionally not checking whether there is an active context
+        // from outside of nextjs to ensure that we can provide the same level
+        // of telemetry when using a custom server
+        const existingRootSpanId = spanContext.getValue(rootSpanIdKey);
+        const isRootSpan = typeof existingRootSpanId !== 'number' || !rootSpanAttributesStore.has(existingRootSpanId);
         const spanId = getSpanId();
         options.attributes = {
             'next.span_name': spanName,
@@ -4303,11 +4305,17 @@ class NextTracerImpl {
             ...options.attributes
         };
         return context.with(spanContext.setValue(rootSpanIdKey, spanId), ()=>this.getTracerInstance().startActiveSpan(spanName, options, (span)=>{
-                const startTime = 'performance' in globalThis && 'measure' in performance ? globalThis.performance.now() : undefined;
+                let startTime;
+                if (NEXT_OTEL_PERFORMANCE_PREFIX && type && __TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$server$2f$lib$2f$trace$2f$constants$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["LogSpanAllowList"].has(type)) {
+                    startTime = 'performance' in globalThis && 'measure' in performance ? globalThis.performance.now() : undefined;
+                }
+                let cleanedUp = false;
                 const onCleanup = ()=>{
+                    if (cleanedUp) return;
+                    cleanedUp = true;
                     rootSpanAttributesStore.delete(spanId);
-                    if (startTime && process.env.NEXT_OTEL_PERFORMANCE_PREFIX && __TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$server$2f$lib$2f$trace$2f$constants$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["LogSpanAllowList"].includes(type || '')) {
-                        performance.measure(`${process.env.NEXT_OTEL_PERFORMANCE_PREFIX}:next-${(type.split('.').pop() || '').replace(/[A-Z]/g, (match)=>'-' + match.toLowerCase())}`, {
+                    if (startTime) {
+                        performance.measure(`${NEXT_OTEL_PERFORMANCE_PREFIX}:next-${(type.split('.').pop() || '').replace(/[A-Z]/g, (match)=>'-' + match.toLowerCase())}`, {
                             start: startTime,
                             end: performance.now()
                         });
@@ -4316,10 +4324,17 @@ class NextTracerImpl {
                 if (isRootSpan) {
                     rootSpanAttributesStore.set(spanId, new Map(Object.entries(options.attributes ?? {})));
                 }
-                try {
-                    if (fn.length > 1) {
+                if (fn.length > 1) {
+                    try {
                         return fn(span, (err)=>closeSpanWithError(span, err));
+                    } catch (err) {
+                        closeSpanWithError(span, err);
+                        throw err;
+                    } finally{
+                        onCleanup();
                     }
+                }
+                try {
                     const result = fn(span);
                     if ((0, __TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$shared$2f$lib$2f$is$2d$thenable$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["isThenable"])(result)) {
                         // If there's error make sure it throws
@@ -4351,7 +4366,7 @@ class NextTracerImpl {
             {},
             args[1]
         ];
-        if (!__TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$server$2f$lib$2f$trace$2f$constants$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["NextVanillaSpanAllowlist"].includes(name) && process.env.NEXT_OTEL_VERBOSE !== '1') {
+        if (!__TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist$2f$esm$2f$server$2f$lib$2f$trace$2f$constants$2e$js__$5b$middleware$2d$edge$5d$__$28$ecmascript$29$__["NextVanillaSpanAllowlist"].has(name) && process.env.NEXT_OTEL_VERBOSE !== '1') {
             return fn;
         }
         return function() {
@@ -10348,7 +10363,7 @@ var __TURBOPACK__imported__module__$5b$project$5d2f$node_modules$2f$next$2f$dist
     exports.useMemo = function(create, deps) {
         return resolveDispatcher().useMemo(create, deps);
     };
-    exports.version = "19.3.0-canary-2bcbf254-20251020";
+    exports.version = "19.3.0-canary-52684925-20251110";
 }();
 }),
 "[project]/node_modules/next/dist/compiled/react/react.react-server.js [middleware-edge] (ecmascript)", ((__turbopack_context__, module, exports) => {
